fix: sticky table headers were transparent in dark mode
- Add a dedicated --bg-card-solid theme variable, opaque in both light and dark mode, separate from the intentionally translucent --bg-card
- Switch sticky thead th backgrounds on Employees, Ticketing, Asset Management, and Email Accounts tables to --bg-card-solid, fixing scrolled content showing through the header in dark mode

+++ STLAF.Api/Departments/HRAdmin/Entities/LeaveType.cs
@@ -5,6 +5,6 @@ namespace STLAF.Api.Departments.HRAdmin.Entities;
 public class LeaveType : BaseEntity
 {
     public string Name { get; set; } = string.Empty;
-    public int DefaultCredits { get; set; }
+    public decimal DefaultCredits { get; set; }
     public int? RequiresMedicalAfterDays { get; set; }
 }
